requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-login==0.6.3
flask-wtf==1.2.1
flask-mail==0.10.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
anthropic==0.28.0
wtforms==3.1.2
werkzeug==3.0.3
sqlalchemy==2.0.31
geoalchemy2==0.15.2
shapely==2.0.5
requests==2.32.3
apscheduler==3.10.4
stripe==10.4.0
pillow==10.4.0
gunicorn==22.0.0
email-validator==2.2.0
